Remove some unused code and fix issues

+++ player/core/src/main/java/dev/jdtech/jellyfin/utils/system/brightness/BrightnessControl.kt
@@ -3,7 +3,7 @@ package dev.jdtech.jellyfin.utils.system.brightness
 interface BrightnessControl {
     fun getBrightness(): Float
 
-    fun setBrightness(value: Int)
+    fun setBrightness(value: Float)
 
     fun changeBrightnessByRatio(ratio: Float)
